Removes the string munging of extractReducer
The previous version of extractReducer breaks the scope of the reducer.
When it gets evaled there is no guarantee that the variables referenced
in the body of the reducer are in scope.

With the new approach, the reducer is used as a normal function, so it
retains all of the variables that it has closed over.